Removed useless computations.
The plugins-manager loads, now, the plugins once. If a new plugn is installed once the software running, it will cache only the new data from the new plugin.

As soon as a plugin is first used, an object is instantiated with its own id and stored in a dictionary as value, with for key, the name of the plugin.

Tested from Windows10 version: 10.0.14393